run_make_support: move external deps to `external_deps` module
