+++ src/tools/run-make-support/src/env_checked.rs
@@ -1,5 +1,5 @@
-use std::ffi::OsString;
 use std::env;
+use std::ffi::OsString;
 
 #[track_caller]
 #[must_use]
